code update

+++ cpp_windows/main.cpp
@@ -3,31 +3,31 @@
 #include "zkp.h"
 
 int main() {
-    uint64_t p = 181;
-    uint64_t g = 2;
-    uint64_t d = 111213119;
-    uint64_t l = 8;
-    uint64_t n = 5;
-    std::vector<uint64_t> pp;
+    int64_t p = 181;
+    int64_t g = 2;
+    int64_t d = 111213119;
+    int64_t l = 8;
+    int64_t n = 5;
+    std::vector<int64_t> pp;
 
     ZKP lib;
     lib.setup(g, d, l, p, pp); 
 
-    for (uint64_t value : pp) {
+    for (int64_t value : pp) {
         std::cout << value << " ";
     }
     std::cout << std::endl;
 
-    std::vector<std::vector<int>> A, B, C;
+    std::vector<std::vector<int64_t>> A, B, C;
     
     try {
         lib.createMat("instructions.txt", A, B, C, p);
 
         // Print the matrices
-        auto printMatrix = [](const std::vector<std::vector<int>>& matrix, const std::string& name) {
+        auto printMatrix = [](const std::vector<std::vector<int64_t>>& matrix, const std::string& name) {
             std::cout << "Matrix " << name << ":" << std::endl;
             for (const auto& row : matrix) {
-                for (int val : row) {
+                for (int64_t val : row) {
                     std::cout << val << " ";
                 }
                 std::cout << std::endl;
