+++ cpp_windows/zkp.h
@@ -10,8 +10,8 @@
 
 class ZKP {
 public:
-    void setup(uint64_t g, uint64_t d, uint64_t l, uint64_t p, std::vector<uint64_t>& pp);
-    void createMat(const std::string& filename, std::vector<std::vector<int>>& A, std::vector<std::vector<int>>& B, std::vector<std::vector<int>>& C, uint64_t p);
+    void setup(int64_t g, int64_t d, int64_t l, int64_t p, std::vector<int64_t>& pp);
+    void createMat(const std::string& filename, std::vector<std::vector<int64_t>>& A, std::vector<std::vector<int64_t>>& B, std::vector<std::vector<int64_t>>& C, int64_t p);
 };
 
 #endif // ZKP_H
